feat: 패키지 정리 및 테스트 클래스 분리
